added rest of routes
